Working on the main problem of the Game
The problem that target tile can be covered with platforms, denying any enemies from entering. The method im trying, is to pick all of the corner tiles and checking all the reachable tiles from them. If from any corner enemies cant reach the target, that tile will not be a placable tile.